pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dictate"
version = "0.1.0"
description = "Local voice-to-text dictation app for macOS"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Dictate Contributors"},
]
keywords = ["dictation", "voice-to-text", "whisper", "macos"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "rumps>=0.4.0",
    "pynput>=1.7.6",
    "pyperclip>=1.8.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
api = ["openai>=1.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
]

[project.scripts]
dictate = "dictate.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
